requirements.txt
Django==3.0.3
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-password-reset==2.0


Markdown==3.1.1

mistune==0.8.4
Pillow==7.0.0
-e .
